move rewards update loop out of zinnia loop
