+++ lib/zinnia.js
@@ -2,13 +2,9 @@ import { execa } from 'execa'
 import Sentry from '@sentry/node'
 import { installBinaryModule, updateSourceFiles, getBinaryModuleExecutable } from './modules.js'
 import os from 'node:os'
-import { ethers } from 'ethers'
-import fs from 'node:fs/promises'
-import { fileURLToPath } from 'node:url'
 import pRetry from 'p-retry'
 import timers from 'node:timers/promises'
 import { join } from 'node:path'
-import * as Name from 'w3name'
 
 const ZINNIA_DIST_TAG = 'v0.19.1'
 const ZINNIA_MODULES = [
@@ -22,9 +18,7 @@ const ZINNIA_MODULES = [
 ]
 const {
   TARGET_ARCH = os.arch(),
-  MODULE_FILTER = '',
-  // https://github.com/filecoin-station/contract-addresses
-  CONTRACT_ADDRESSES_IPNS_KEY = 'k51qzi5uqu5dmaqrefqazad0ca8b24fb79zlacfjw2awdt5gjf2cr6jto5jyqe'
+  MODULE_FILTER = ''
 } = process.env
 
 export const install = () => installBinaryModule({
@@ -72,9 +66,6 @@ const maybeReportErrorToSentry = (/** @type {unknown} */ err) => {
   Sentry.captureException(err, hint)
 }
 
-let lastTotalJobsCompleted = 0
-let lastRewardsScheduledForAddress = 0n
-
 const matchesModuleFilter = module =>
   MODULE_FILTER === '' || module === MODULE_FILTER
 
@@ -120,35 +111,6 @@ const updateAllSourceFiles = async ({
   return hasUpdated
 }
 
-async function getContractAddresses () {
-  const name = Name.parse(CONTRACT_ADDRESSES_IPNS_KEY)
-  const revision = await Name.resolve(name)
-  return revision.value.split('\n').filter(Boolean)
-}
-
-async function getContractsWithRetry ({ provider, abi, signal }) {
-  const contractAddresses = await pRetry(getContractAddresses, {
-    signal,
-    retries: 10,
-    onFailedAttempt: err => {
-      console.error(err)
-      console.error('Failed to get contract addresses. Retrying...')
-      if (String(err).includes('You are being rate limited')) {
-        const delaySeconds = 60 + (Math.random() * 60)
-        // Don't DDOS the w3name services
-        console.error(
-          `Rate limited. Waiting ${delaySeconds} seconds...`
-        )
-        return timers.setTimeout(delaySeconds * 1000)
-      }
-    }
-  })
-  console.error(`Meridian contract addresses: ${contractAddresses.join(', ')}`)
-  return contractAddresses.map(address => {
-    return new ethers.Contract(address, abi, provider)
-  })
-}
-
 const runUpdateSourceFilesLoop = async ({
   controller,
   signal,
@@ -189,51 +151,6 @@ const runUpdateSourceFilesLoop = async ({
       })
       console.error(err)
       maybeReportErrorToSentry(err)
-    }
-  }
-}
-
-const runUpdateContractsLoop = async ({ signal, provider, abi, contracts }) => {
-  while (true) {
-    const delay = 10 * 60 * 1000 // 10 minutes
-    const jitter = Math.random() * 20_000 - 10_000 // +- 10 seconds
-    try {
-      await timers.setTimeout(delay + jitter, null, { signal })
-    } catch (err) {
-      if (err.name === 'AbortError') return
-      throw err
-    }
-    const newContracts = await getContractsWithRetry({ provider, abi, signal })
-    contracts.splice(0)
-    contracts.push(...newContracts)
-    if (signal.aborted) {
-      return
-    }
-  }
-}
-
-const runUpdateRewardsLoop = async ({ signal, contracts, ethAddress, onMetrics }) => {
-  while (true) {
-    const contractRewards = await Promise.all(contracts.map(async contract => {
-      return getScheduledRewardsWithFallback(contract, ethAddress)
-    }))
-    if (signal.aborted) {
-      return
-    }
-    const totalRewards = contractRewards.reduce((a, b) => a + b, 0n)
-    onMetrics({
-      totalJobsCompleted: lastTotalJobsCompleted,
-      rewardsScheduledForAddress: totalRewards
-    })
-    lastRewardsScheduledForAddress = totalRewards
-
-    const delay = 10 * 60 * 1000 // 10 minutes
-    const jitter = Math.random() * 20_000 - 10_000 // +- 10 seconds
-    try {
-      await timers.setTimeout(delay + jitter, null, { signal })
-    } catch (err) {
-      if (err.name === 'AbortError') return
-      throw err
     }
   }
 }
@@ -288,6 +205,8 @@ const catchChildProcessExit = async ({
 }
 
 export async function run ({
+  provider,
+  abi,
   STATION_ID,
   FIL_WALLET_ADDRESS,
   ethAddress,
@@ -297,28 +216,10 @@ export async function run ({
   moduleSourcesDir,
   onActivity,
   onMetrics,
-  isUpdated = false
+  isUpdated = false,
+  lastTotalJobsCompleted,
+  lastRewardsScheduledForAddress
 }) {
-  const fetchRequest = new ethers.FetchRequest(
-    'https://api.node.glif.io/rpc/v1'
-  )
-  fetchRequest.setHeader(
-    'Authorization',
-    'Bearer RXQ2SKH/BVuwN7wisZh3b5uXStGPj1JQIrIWD+rxF0Y='
-  )
-  const provider = new ethers.JsonRpcProvider(
-    fetchRequest,
-    null,
-    { batchMaxCount: 1 }
-  )
-  const abi = JSON.parse(
-    await fs.readFile(
-      fileURLToPath(new URL('./abi.json', import.meta.url)),
-      'utf8'
-    )
-  )
-
-  const contracts = await getContractsWithRetry({ provider, abi, signal: null })
   const zinniadExe = getBinaryModuleExecutable({ module: 'zinnia', executable: 'zinniad' })
 
   if (!isUpdated) {
@@ -388,11 +289,11 @@ export async function run ({
       resetTimeout()
       handleEvents({
         module,
-        contracts,
-        ethAddress,
         onActivity,
         onMetrics,
-        text: data
+        text: data,
+        lastTotalJobsCompleted,
+        lastRewardsScheduledForAddress
       }).catch(err => {
         console.error(err)
         Sentry.captureException(err)
@@ -420,8 +321,6 @@ export async function run ({
         moduleVersionsDir,
         moduleSourcesDir
       }),
-      runUpdateContractsLoop({ signal, provider, abi, contracts }),
-      runUpdateRewardsLoop({ signal, contracts, ethAddress, onMetrics }),
       catchChildProcessExit({ childProcesses, onActivity, controller })
     ])
     console.error('Zinnia main loop ended')
@@ -435,6 +334,8 @@ export async function run ({
   // This infinite recursion has no risk of exceeding the maximum call stack
   // size, as awaiting promises unwinds the stack
   return run({
+    provider,
+    abi,
     STATION_ID,
     FIL_WALLET_ADDRESS,
     ethAddress,
@@ -444,7 +345,9 @@ export async function run ({
     moduleSourcesDir,
     onActivity,
     onMetrics,
-    isUpdated: true
+    isUpdated: true,
+    lastTotalJobsCompleted,
+    lastRewardsScheduledForAddress
   })
 }
 
@@ -452,11 +355,11 @@ const jobsCompleted = {}
 
 async function handleEvents ({
   module,
-  contracts,
-  ethAddress,
   onActivity,
   onMetrics,
-  text
+  text,
+  lastTotalJobsCompleted,
+  lastRewardsScheduledForAddress
 }) {
   for (const line of text.trimEnd().split(/\n/g)) {
     try {
@@ -492,9 +395,9 @@ async function handleEvents ({
           const totalJobsCompleted = Object.values(jobsCompleted).reduce((a, b) => a + b, 0)
           onMetrics({
             totalJobsCompleted,
-            rewardsScheduledForAddress: lastRewardsScheduledForAddress
+            rewardsScheduledForAddress: lastRewardsScheduledForAddress.get()
           })
-          lastTotalJobsCompleted = totalJobsCompleted
+          lastTotalJobsCompleted.set(totalJobsCompleted)
           break
         }
 
@@ -504,14 +407,5 @@ async function handleEvents ({
     } catch (err) {
       console.error('Ignoring malformed Zinnia event:', line)
     }
-  }
-}
-
-async function getScheduledRewardsWithFallback (contract, ethAddress) {
-  try {
-    return await contract.rewardsScheduledFor(ethAddress)
-  } catch (err) {
-    console.error('Failed to get scheduled rewards:', err.stack)
-    return 0n
   }
 }
